refactor: change folder and files name convention
